+++ src/lerp/index.ts
@@ -1,0 +1,1 @@
+export { createLerp } from "./create-lerp";